Move files to new nomenclature
